+++ installer/templates/new/config/prod.exs
@@ -1,0 +1,6 @@
+use Mix.Config
+
+config :kitto, reload_code?: false, watch_assets?: false, serve_assets?: true
+
+# For distillery releases
+# config :kitto, root: :otp_app