+++ lib/kitto.ex
@@ -21,7 +21,7 @@ defmodule Kitto do
   def start(_type, _args) do
     opts = [strategy: :one_for_one, name: Kitto.Supervisor]
 
-    Supervisor.start_link(children(Mix.env), opts)
+    Supervisor.start_link(children(), opts)
   end
 
   def start_server do
@@ -34,6 +34,7 @@ defmodule Kitto do
   """
   def root do
     case Application.get_env(:kitto, :root) do
+      :otp_app -> Application.app_dir(Application.get_env(:kitto, :otp_app))
       path when is_bitstring(path) -> path
       nil ->
         """
@@ -43,6 +44,11 @@ defmodule Kitto do
         exit(:shutdown)
     end
   end
+
+  @doc """
+  Returns true when the asset development server is set to be watching for changes
+  """
+  def watch_assets?, do: Application.get_env :kitto, :watch_assets?, true
 
   @doc """
   Returns the binding ip of the assets watcher server
@@ -75,14 +81,15 @@ defmodule Kitto do
   defp port(p) when is_integer(p), do: p
   defp port(_), do: @defaults.port
 
-  defp children(:dev) do
+
+  defp children do
     case Kitto.CodeReloader.reload_code? do
-      true -> children(:all) ++ [worker(Kitto.CodeReloader, [[server: :runner]])]
-      false -> children(:all)
+      true -> children(:prod) ++ [worker(Kitto.CodeReloader, [[server: :runner]])]
+      false -> children(:prod)
     end
   end
 
-  defp children(_env) do
+  defp children(:prod) do
     [supervisor(__MODULE__, [], function: :start_server),
      supervisor(Kitto.Notifier, []),
      worker(Kitto.BackoffServer, [[]]),