+++ lib/kitto/job/dsl.ex
@@ -54,6 +54,7 @@ defmodule Kitto.Job.DSL do
     broadcast events using the jobs name.
   """
   defmacro job(name, options, contents \\ []) do
+    name = if is_atom(name), do: name, else: String.to_atom(name)
     if options[:command] do
       _job(:shell, name, options)
     else
@@ -63,7 +64,8 @@ defmodule Kitto.Job.DSL do
 
   defp _job(:elixir, name, options, contents) do
     block = Macro.prewalk (options[:do] || contents[:do]), fn
-      {:broadcast!, meta, args = [_]} -> {:broadcast!, meta, [name] ++ args}
+      {:|>, pipe_meta, [lhs, {:broadcast!, meta, context}]} when is_atom(context) or context == [] -> {:|>, pipe_meta, [lhs, {:broadcast!, meta, [name]}]} 
+      {:broadcast!, meta, args = [{:%{}, _, _}]} -> {:broadcast!, meta, [name] ++ args}
       ast_node -> ast_node
     end
 
@@ -86,7 +88,7 @@ defmodule Kitto.Job.DSL do
         Notifier.broadcast!(unquote(name), %{stdout: stdout, exit_code: exit_code})
       end
 
-      Job.register binding[:runner_server],
+      Job.register binding()[:runner_server],
                    unquote(name),
                    unquote(options),
                    (__ENV__ |> Map.take([:file, :line])),